+++ src/adkcore.h
@@ -8,6 +8,8 @@
 
 namespace adk
 {
+    bool add_global_callback(void (*func)(void));
+
     // run ardukit engine: task queue and timer events
     void run();