+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Optimize timer memory footprint
 - Add timer event queue resizng support
 - Add custom scan interval for Button and EdgeDetector
+- Add add_global_callback() to adkcore to reduce dependency to timer and task modules
 
 ## 1.1
 - Add EdgeDetector (deprecating EdgeTrigger)